elimine casi todos los warnings

+++ SFML 2.3.21/Block.cpp
@@ -4,7 +4,7 @@
 Block::Block(const Resources& res, const b2Vec2& size, const b2Vec2& position):
 	mSize(size),
 	mSprite(res.textures.get(Texture::SPRITE_BLOCK),
-		sf::IntRect(0,0,size.x*PIXELS_PER_METER,size.y*PIXELS_PER_METER)
+		sf::IntRect(0,0,static_cast<int>(size.x*PIXELS_PER_METER),static_cast<int>(size.y*PIXELS_PER_METER))
 	)
 {
 	auto bounds=mSprite.getLocalBounds();
